fix(beta): closed-beta triage — int64 overflow, gem economy, quest exploit, raid/stat fixes
Backend batch from the closed-beta playtest triage (1005 unit tests green).

Economy / overflow:
- Widen gem ledger + stat/reward fields int->long/bigint (no late-game overflow).
  Migrations: WidenGemAmountToBigint, WidenStatAndRewardFieldsToBigint.
- Boss gems: flat 2 on a per-chapter-scaled chance, unified across quest + raid bosses;
  granted only on boss clear (removes per-hit gem inflation).

Quests:
- Node depletion now keyed per (player, quest, difficulty) — closes the cross-difficulty
  sigil exploit. Migration: AddQuestProgressDifficulty (backfill Normal).
- Per-difficulty progress ladder; boss gem gated to nodeJustCleared.

Raids:
- Deterministic sort (low-HP first); per-hit health drain surfaced in the hit response;
  tier label from the raid definition; guild-stamina pool reported correctly; completed-tab
  recency window; loot = dismiss-on-last-claim.

Stats:
- LSI cap 7.45; stat-alloc per-call cap 100 -> 100M; AllocateStatResponse carries
  Effective ATK/DEF for inline chip refresh.

Content / infra:
- Remove inflated on-kill XP bonus (throughput model); Pano set 15/15 + steed proc;
  XP-to-next-level on profile; response compression; magic catalogue endpoint.

+++ src/ROTA.Application/Interfaces/IEquipmentService.cs
@@ -10,7 +10,7 @@ public interface IEquipmentService
     Task<IReadOnlyList<OwnedGearResponse>>    GetOwnedGearAsync(Guid playerId, CancellationToken ct = default);
 
     // Called by RaidService on every hit. baseAtk/baseDef are from PlayerStats.
-    Task<EffectiveCombatData> GetEffectiveCombatDataAsync(Guid playerId, int baseAtk, int baseDef, CancellationToken ct = default);
+    Task<EffectiveCombatData> GetEffectiveCombatDataAsync(Guid playerId, long baseAtk, long baseDef, CancellationToken ct = default);
 
     /// <summary>
     /// Idempotent gear grant — upsert: adds to existing stack or creates a new row.
@@ -21,8 +21,8 @@ public interface IEquipmentService
 
 // Lives in this file alongside the interface.
 public sealed record EffectiveCombatData(
-    int           EffectiveAttack,
-    int           EffectiveDefense,
+    long          EffectiveAttack,
+    long          EffectiveDefense,
     GearProcData? MountProc,           // null when no mount is equipped
     double        FlatDamagePercent);  // accumulated from conditional bonuses; 0.0 when none
